requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Script==2.0.6
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
newsapi-python==0.2.6
pymongo==3.10.1
requests==2.22.0
urllib3==1.25.8
Werkzeug==0.16.0
